# Replace hand-written endian-specific loads with std.mem.readInt*()