And when we have the choice, favor little-endian because it's 2023.

Gives a slight performance improvement:

```
   md5: 552 -> 555 MiB/s
  sha1: 768 -> 786 MiB/s
sha512: 211 -> 217 MiB/s
```